Merge player list image parsing

+++ cogs/Raiding/vc_select.py
@@ -6,16 +6,17 @@ import sql
 
 
 class VCSelect:
-    def __init__(self, client, ctx, headcount=False, lock=False, unlock=False, clean=False):
+    def __init__(self, client, ctx, headcount=False, lock=False, unlock=False, clean=False, parse=False):
         self.client = client
         self.ctx = ctx
         self.headcount = headcount
+        self.parse = parse
         title = "Headcount Setup" if headcount else "Lock Selection" if lock else "Unlock Selection" if unlock else "Cleaning Selection" if\
-            clean else "AFK-Check Setup"
+            clean else "Parsing Setup" if parse else "AFK-Check Setup"
         description = "Please choose what channel you'd like to start this headcount in." if headcount else\
             "Please choose which channel you'd like to lock." if lock else "Please choose which channel you'd like to unlock." if unlock\
-            else "Please choose which channel you'd like to clean." if clean else \
-                "Please choose what channel you'd like to start this afk check in."
+            else "Please choose which channel you'd like to clean." if clean else "Please choose which channel you'd like to parse for." if\
+                parse else "Please choose what channel you'd like to start this afk check in."
         self.locationembed = discord.Embed(title=title,
                                            description=description,
                                            color=discord.Color.green())
@@ -26,7 +27,8 @@ class VCSelect:
 
 
     async def start(self):
-        await self.ctx.message.delete()
+        if not self.parse:
+            await self.ctx.message.delete()
 
         if self.ctx.channel == self.guild_db.get(sql.gld_cols.raidcommandschannel):
             s = ""
